cmake: add initial structure
This structure illustrates precisely the type of separation one should
strive for; clear, unambiguous and readily identifiable components.

This tutorial project officially only supports two compilers: g++ >=8.1,
or clang++ >=5.0. If neither compilers are used, a warning will be
emitted to the user but it is not an error. It is important to note that
failure to use either one of these compilers may result in poor code
generation (and therefore, poor performance) or a failed compile
attempt. This tutorial project does not cover these scenarios.

This tutorial project is written in C++17 with support for nonstandard
extensions which is a partial violation of rule "P.2: Write in ISO
Standard C++" of the C++ Core Guidelines. We violate it because we
specifically only support two compilers, of which they are ubiquitous
and accessible on almost every system. In addition, exceptions and RTTI
are forbidden throughout the project and are forcibly disabled through
compiler options if supported.

Once a target has been created (e.g. through an `add_library()` call),
it is imperative that one calls `vmtutorial_configure_target()` to set
build flags on it relative to the specified build type, and to enforce
the aforementioned.