core: add disassembler
We need to be able to see what's going on as soon as possible, so a
disassembler is wise. A disassembler will allow us to translate an
instruction value into the equivalent CHIP-8 human-readable assembly
language.

Some minor changes to the build system had to be made to accomplish
this, but they aren't noteworthy changes.

This commit provides a test case to verify that the disassembler works
correctly using googletest as the backend.