implementation: add interpreter
The interpreter implementation is your standard fetch-decode-execute
loop. It ships with a unit test which can be used for future
implementations with hopefully an addition of an implementation class to
the `ImplementationTypes` type alias. However, the unit test could use
some cleaning up and thus it is not extremely documented as it is likely
subject to change. The individual unit tests work just fine, though.

We've removed the cognitive complexity warning from .clang-tidy, as
that's drug-induced garbage. The amount of "NOLINT" and the like
littered throughout the code is a huge eye sore, and I'm not sure if I'm
doing something wrong or if that's the Right Way (TM). I have no idea.

All of the paragraphs upon paragraphs of comments pertaining to
technical details should be treated as rough drafts and *not* final. Be
gentle.