interpreter: change names and add methods
This commit changes the key press handling methods to say "halt" for the
most part, to be consistent with the naming of other methods.

This commit adds a `GetPlaceValues()` method which will return the
digits at the hundreds, tens, and ones places. This is purely a
convenience function and takes away the necessity to do the same math
over and over again among different implementations.