frontend: Add user interface
This user interface is based on Qt, however it is in progress and it
lacks features, such as playing tones or user preferences, or even a
debugger. Keyboard input is also not recognized yet.

We're not going to implement sounds until Qt 6.2, at which time Qt
Multimedia should be released. Documentation needs work.

At any rate, guest programs do work and operate visually fine so far.