frontend: add controllers and models
We add an AppSettingsModel to handle configuration settings. This is
not a model like you'd expect a la Qt's model/view architecture,
although I suspect it could be in the future. Nevertheless per the MVC
paradigm, it is a kind of data handler, so it's still by any means a
model.

We've added controllers for the settings dialogs, and the keypad
settings handle bindings. It's a bit crummy, but it works right now,
and can always be improved.

Fixed a bug within the thread handler where we used `QThread::quit()`
instead of `QThread::requestInterruption()` to halt execution of the
VM thread.

Folder structure was reorganized.