frontend: add icons and FPS counter
We've added toolbar icons to make the interface look more spiffy.
The icons are of somewhat poor resolution, so they will probably be
changed in the future.

The title bar will now display the running guest program.

The FPS counter will be displayed in the status bar at the right hand
corner.