frontend: fix frame limiter and graphics
For some very strange reason, an extra `glBindTexture()` function call
was necessary to get OpenGL to play nice with Qt. Prior to this change,
resizing the window was the only way to actually display the updates to
the graphics. I don't have a solid explanation for this, yet. All I know
is that this works with no obvious ill effects. I'm not even sure this
is an optimal solution...

The frame limiter had a serious flaw on Windows. The granularity of
the `Sleep()` function on Windows (which `QThread::msleep()` uses) is
~16ms, which is sad news because 60FPS is ~16ms. Generally, CHIP-8
frames take (or rather should take on average) 0ms and the default FPS
target is, you guessed it, ~16ms. Executing the original code on Windows
causes the frame limiter to slow the virtual machine to half of the
requested FPS.

The new code makes use of a deadline concept using the native C++
threading facilities.

On a slightly unrelated note, we should use as much of the C++ standard
library as possible, as that should be taking precedence over a number
of Qt's duplicated facilities...