frontend/core: fix audio
The core wasn't returning milliseconds for the audio duration: fixed.

The frontend code used for generating and outputting the audio samples
was mostly taken from the Qt Multimedia examples, however it seems to be
way too complicated for our purposes. We just directly write the buffer
to the audio output now without complexity.