project: documentation, core frontend update
We use doxygen-awesome-css for the theme when generating the code
documentation. This is added as a submodule.

The core now contains a very basic logger using libfmt to handle the
format strings.

The audio, general, and machine settings now operate properly.

There is still much to be done in the way of documentation and features.