frontend: add square/triange/sawtooth waves, fixes
The choice of audio device and tone is now correctly saved within the
configuration file.

Moved from QtMultimedia to SDL2. This is because in all of my tests with
it, the behavior was inconsistent among operating systems. Plus, SDL is
well known unlike QtMultimedia's rework, which still needs time for
people to become acquainted with it. I wanted to use SDL before, but I
didn't want to introduce another dependency. It however, could no longer
be helped.

The settings dialog is now modal. You will not be able to manipulate
other windows within the program while the settings dialog exists.

The virtual machine manager (VMThread) correctly sets itself up from
configuration settings.

The only thing that's missing from this commit is proper audio volume
management.