frontend/core: Fixes and cleanups
Make step into work like every other debugger in existence. I purposely
wanted to not do that in this case to emphasize scoping, but it would be
incorrect to the point of basically telling a lie.

Remove the logger methods from the VMInstance class as they're
superfluous since the Logger class is now exposed

Add a `FindBreakpoint()` method for finding breakpoints. Returns an iterator.

Use breakpoint and pointer icons in the disassembly view.